+++ delivery-service/src/test/java/com/chill_guys/delivery_service/DeliveryServiceImplApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DeliveryServiceApplicationTests {
+class DeliveryServiceImplApplicationTests {
 
 	@Test
 	void contextLoads() {
